Don't strip whitespace
Stripping whitespace breaks when &commentstring contains a space but not
all comments strictly adhere to &commentstring format.

For example, in Perl, where &commentstring is "# %s", commenting this
block

  # Normal comment
  #
  3

ought to produce

  # # Normal comment
  # #
  # 3

but instead produces

  ## Normal comment
  ##
  #3

This breaks *toggle* functionality, because ## lines are not recognized
as comments, and so the next invocation adds *another* #:

  ### Normal comment
  ###
  ##3

I don't know what cases the s:strip_white_space() call is supposed to
help with, so I don't know what removing it breaks (i.e., there are no
tests).